chore: Move different model types to dedicate files
