+++ AWSAuthSDK/Sources/AWSMobileClient/Models/SignInResult.swift
@@ -1,0 +1,28 @@
+//
+// Copyright 2017-2018 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License").
+// You may not use this file except in compliance with the License.
+// A copy of the License is located at
+//
+// http://aws.amazon.com/apache2.0
+//
+// or in the "license" file accompanying this file. This file is distributed
+// on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+// express or implied. See the License for the specific language governing
+// permissions and limitations under the License.
+
+import Foundation
+
+/// Contains the result of the sign in operation.
+public struct SignInResult {
+    public let signInState: SignInState
+    public let parameters: [String: String]
+    public let codeDetails: UserCodeDeliveryDetails?
+
+    internal init(signInState: SignInState, parameters: [String: String] = [:], codeDetails: UserCodeDeliveryDetails? = nil) {
+        self.signInState = signInState
+        self.parameters = parameters
+        self.codeDetails = codeDetails
+    }
+}
